Add toy test as initial demo of using WebDriver w/ LambdaTest
This is like the toy test as the previous one written to use the
Firefox WebDriver, but this time using the WebDriver to connect to
LambdaTest through the LambdaTest tunnel.

Running this test results in the test being run on the LambdaTest
Selenium Grid infrastructure.

*Caveat* My access key is embedded in LambdaTests.cs. This is a bad
practice for security, but I think probably is acceptable for the
school assignment.